+++ extensions/bitbucket/src/components/pullRequests/searchAllPullRequests.tsx
@@ -1,4 +1,4 @@
-import { ActionPanel, List, showToast, Color, Action, Icon, Image, Toast } from "@raycast/api";
+import { ActionPanel, List, showToast, Color, Action, Icon, Image, Toast, getPreferenceValues } from "@raycast/api";
 import { useEffect, useState } from "react";
 import useSWR, { SWRConfig } from "swr";
 
@@ -9,18 +9,20 @@ import {
   RequestChangesAction,
   ShowPullRequestDetailAction,
 } from "./actions";
-import { cacheConfig } from "../../helpers/cache";
+import { scanCacheConfig } from "../../helpers/cache";
 import { preferences } from "../../helpers/preferences";
 import { PullRequest } from "./interface";
 import { getPullRequestKey } from "./../../helpers/pullRequestKey";
 import { ReviewState, setReviewState } from "./../../helpers/reviewState";
 import { buildReviewAccessories, findMyReviewState } from "./../../helpers/reviewers";
 
-const PULL_REQUESTS_CACHE_KEY = `all-open-pull-requests:${preferences.workspace}:${preferences.email}`;
+const commandPreferences = getPreferenceValues<Preferences.SearchAllPullRequests>();
+
+const PULL_REQUESTS_CACHE_KEY = `all-open-pull-requests:${preferences.workspace}:${preferences.email}:${commandPreferences.maxRepoAgeDays || "0"}`;
 
 export function SearchAllPullRequests() {
   return (
-    <SWRConfig value={cacheConfig}>
+    <SWRConfig value={scanCacheConfig}>
       <SearchAllPullRequestsList />
     </SWRConfig>
   );