+++ extensions/bitbucket/src/helpers/preferences.ts
@@ -1,9 +1,3 @@
 import { getPreferenceValues } from "@raycast/api";
 
-export const preferences: Preferences = getPreferenceValues();
-
-export interface Preferences {
-  workspace: string;
-  email: string;
-  apiToken: string;
-}
+export const preferences = getPreferenceValues<Preferences>();